# server/docs/swagger.yaml
definitions:
  ds.Book:
    properties:
      author_link:
        type: string
      author_name:
        type: string
      cover_file_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      homepage:
        type: string
      id:
        type: string
      owner_id:
        type: string
      preview_file_id:
        type: string
      public_id:
        type: string
      published_at:
        type: string
      release_date:
        type: string
      status:
        $ref: '#/definitions/ds.EntityStatus'
      title:
        type: string
      updated_at:
        type: string
      visibility:
        $ref: '#/definitions/ds.EntityVisibility'
    type: object
  ds.EntityStatus:
    enum:
    - review
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - EntityStatusUnderReview
    - EntityStatusApproved
    - EntityStatusRejected
  ds.EntityVisibility:
    enum:
    - public
    - private
    - unlisted
    type: string
    x-enum-varnames:
    - EntityVisibilityPublic
    - EntityVisibilityPrivate
    - EntityVisibilityUnlisted
  ds.File:
    properties:
      created_at:
        type: string
      id:
        type: string
      mime_type:
        type: string
      name:
        type: string
      owner_id:
        type: string
      purpose:
        $ref: '#/definitions/ds.FilePurpose'
      size:
        type: integer
      type:
        $ref: '#/definitions/file.Type'
    type: object
  ds.FilePurpose:
    enum:
    - book-cover
    type: string
    x-enum-varnames:
    - FilePurposeBookCover
  file.Type:
    enum:
    - other
    - image
    - video
    type: string
    x-enum-varnames:
    - TypeOther
    - TypeImage
    - TypeVideo
  handler.Error:
    properties:
      code:
        type: integer
      error:
        type: string
      errors:
        items:
          type: string
        type: array
      input_errors:
        additionalProperties:
          type: string
        type: object
    type: object
...
  request.CreateBook:
    properties:
      author_link:
        type: string
      author_name:
        type: string
      cover_file_id:
        type: string
      description:
        type: string
      homepage:
        type: string
      release_date:
        type: string
      title:
        type: string
      visibility:
        $ref: '#/definitions/ds.EntityVisibility'
    type: object
...
  response.FilterBooks:
    properties:
      count:
        type: integer
      data:
        items:
          $ref: '#/definitions/ds.Book'
        type: array
    type: object
  response.Status:
    properties:
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /books/:
    get:
      consumes:
      - application/json
      operationId: FilterBooks
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: per_page
        type: integer
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.FilterBooks'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Filter books
      tags:
      - books
    post:
      consumes:
      - application/json
      operationId: CreateBook
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateBook'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ds.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Create a new book
      tags:
      - books
  /books/{id}/:
    get:
      consumes:
      - application/json
      operationId: GetBook
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ds.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Get book
      tags:
      - books
  /files/:
    post:
      consumes:
      - multipart/form-data
      operationId: UploadFile
      parameters:
      - description: File
        in: body
        name: file
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: File purpose
        in: body
        name: purpose
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Status'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Upload file
      tags:
      - files
  /files/{id}/:
    delete:
      consumes:
      - application/json
      operationId: DeleteFile
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Status'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete file
      tags:
      - files
    post:
      consumes:
      - application/json
      operationId: GetFileMetadata
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.File'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Get file metadata
      tags:
      - files
...
swagger: "2.0"